refactor: 패키지 구조 변경, 각각의 도메인을 구분
