+++ src/main/java/racingcar/domain/car/CarFactory.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import racingcar.exception.CarValidator;
