+++ src/main/java/racingcar/domain/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
